project_guardian/requirements.txt
# Core dependencies
openai>=1.0.0
flask>=2.0.0
flask-socketio>=5.3.0
typing-extensions>=4.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
pyttsx3>=2.90

# Enhanced Project Guardian Learning & Memory Dependencies
aiohttp>=3.8.0
feedparser>=6.0.0
sqlite3>=3.35.0
asyncio>=3.4.3
dataclasses>=0.6
pathlib>=1.0.1
uuid>=1.30
hashlib>=3.2

# Web Scraping and Data Processing
lxml>=4.6.0
html5lib>=1.1
cssselect>=1.1.0
chardet>=4.0.0

# Database and Storage
sqlite3>=3.35.0
json5>=0.9.0

# Async and Concurrency
aiofiles>=0.8.0
asyncio-mqtt>=0.11.0

# Logging and Monitoring
structlog>=21.5.0
colorlog>=6.7.0

# Data Processing and Validation
pydantic>=1.10.0
marshmallow>=3.19.0

# Testing and Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Code Quality
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Additional dependencies for enhanced features
pyyaml>=5.4.0
psutil>=5.8.0

# Vector Memory and Embeddings
faiss-cpu>=1.7.4
numpy>=1.21.0

# Security and Cryptography
cryptography>=3.4.0
bcrypt>=4.0.0
PyJWT>=2.6.0

# Performance and Monitoring
prometheus-client>=0.15.0
statsd>=4.0.0

# Development Tools
pre-commit>=2.20.0
tox>=4.0.0 